HSKLevel should live with Chinese definitions
