+++ api-v2/app/com/foreignlanguagereader/api/domain/definition/combined/ChineseDefinition.scala
@@ -1,10 +1,12 @@
 package com.foreignlanguagereader.api.domain.definition.combined
 
-import com.foreignlanguagereader.api.HSKLevel.HSKLevel
+import com.foreignlanguagereader.api
 import com.foreignlanguagereader.api.domain.definition.entry.DefinitionSource
 import com.foreignlanguagereader.api.domain.definition.entry.DefinitionSource.DefinitionSource
 import com.foreignlanguagereader.api.dto.v1.definition.ChineseDefinitionDTO
-import com.foreignlanguagereader.api.{HSKLevel, Language}
+import com.foreignlanguagereader.api.Language
+import com.foreignlanguagereader.api.domain.definition.combined.HSKLevel.HSKLevel
+import play.api.libs.json.{Format, JsResult, JsString, JsSuccess, JsValue}
 
 case class ChineseDefinition(override val subdefinitions: List[String],
                              override val tag: String,
@@ -44,4 +46,22 @@ object ChineseDefinition {
     definitions: Seq[ChineseDefinition]
   ): Seq[ChineseDefinitionDTO] =
     definitions.map(x => chineseDefinitionToDefinitionDTO(x))
+}
+
+object HSKLevel extends Enumeration {
+  type HSKLevel = Value
+
+  val ONE: Value = Value("1")
+  val TWO: Value = Value("2")
+  val THREE: Value = Value("3")
+  val FOUR: Value = Value("4")
+  val FIVE: Value = Value("5")
+  val SIX: Value = Value("6")
+  val NONE: Value = Value("")
+
+  implicit val hskLevelFormat: Format[HSKLevel] = new Format[HSKLevel] {
+    def reads(json: JsValue): JsResult[Value] =
+      JsSuccess(HSKLevel.withName(json.as[String]))
+    def writes(level: HSKLevel.HSKLevel) = JsString(level.toString)
+  }
 }
