+++ api-v2/app/com/foreignlanguagereader/api/domain/definition/entry/DefinitionEntry.scala
@@ -3,10 +3,11 @@ package com.foreignlanguagereader.api.domain.definition.entry
 import com.foreignlanguagereader.api.Language.Language
 import com.foreignlanguagereader.api.domain.definition.combined.{
   ChineseDefinition,
-  Definition
+  Definition,
+  HSKLevel
 }
 import com.foreignlanguagereader.api.domain.definition.entry.DefinitionSource.DefinitionSource
-import com.foreignlanguagereader.api.{HSKLevel, Language}
+import com.foreignlanguagereader.api.Language
 import com.sksamuel.elastic4s.{Hit, HitReader}
 import play.api.libs.json._
